refactor: 내 밥모임 페이지의 기존 명령형 코드를 Suspense, ErrorBoundary를 활용해 선언형 코드로 변경
