+++ src/pages/food-party/list/my.tsx
@@ -1,46 +1,27 @@
 import { Flex, Heading } from '@chakra-ui/react';
+import CustomSuspense from 'components/common/CustomSuspense';
+import ErrorBoundary from 'components/common/ErrorBoundary';
 import GoHomeWhenErrorInvoked from 'components/common/GoHomeWhenErrorInvoked';
-import FoodPartyList from 'components/FoodParty/FoodPartyList';
 import FoodPartyListSkeleton from 'components/FoodParty/FoodPartyListSkeleton';
-import { useGetMyFoodPartyList } from 'hooks/query/useFoodParty';
+import MyFoodPartyList from 'components/FoodParty/MyFoodPartyList';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
-import ROUTING_PATHS from 'utils/constants/routingPaths';
 
-const MyFoodPartyList = () => {
-  const router = useRouter();
-  // To Do: 실시간 업데이트를 위한 refetch 필요 by 승준
-  const { data: myFoodPartyList, isLoading, error, isSuccess } = useGetMyFoodPartyList();
-  const handleClickViewFoodPartyButton = (partyId: number) => {
-    router.push(ROUTING_PATHS.FOOD_PARTY.DETAIL.INFORMATION(partyId));
-  };
-  const handleClickReviewFoodPartyButton = (partyId: number, partyName: string) => {
-    router.push(ROUTING_PATHS.FOOD_PARTY.REVIEW(partyId, partyName));
-  };
-
-  if (isLoading) return <FoodPartyListSkeleton foodPartyCount={2} />;
-  if (error) return <GoHomeWhenErrorInvoked />;
-
+const MyFoodPartyListPage = () => {
   return (
     <>
       <Head>
         <title>내가 참여한 밥모임 목록</title>
       </Head>
-      {isSuccess ? (
-        <Flex flexDirection='column' padding='1rem'>
-          <Heading paddingBottom='1rem'>나의 밥모임 목록</Heading>
-          <FoodPartyList
-            isMyFoodParty
-            foodPartyList={myFoodPartyList}
-            onClickViewButton={handleClickViewFoodPartyButton}
-            onClickReviewButton={handleClickReviewFoodPartyButton}
-          />
-        </Flex>
-      ) : (
-        <GoHomeWhenErrorInvoked />
-      )}
+      <Flex flexDirection='column' padding='1rem'>
+        <Heading paddingBottom='1rem'>나의 밥모임 목록</Heading>
+        <ErrorBoundary fallback={<GoHomeWhenErrorInvoked />}>
+          <CustomSuspense fallback={<FoodPartyListSkeleton foodPartyCount={2} />}>
+            <MyFoodPartyList />
+          </CustomSuspense>
+        </ErrorBoundary>
+      </Flex>
     </>
   );
 };
 
-export default MyFoodPartyList;
+export default MyFoodPartyListPage;
